add: Add exception and optional skipping for recursive protobuf messages (#107)
This PR concerns the parsing of protobuf messages with recursive cycles
in its schema definitions, which unfortunately do exist in the wild. The
current behaviour of the package with such messages is to go on an
infinite recursion loop until the eventual `RecursionError: maximum
recursion depth exceeded` crash.

With the proposed changes, we leverage the current structure of the code
to detect cycles by maintaining a trace of the nested messages in the
protobuf schema. Upon encountering a cycle (i.e. a message seen twice on
the trace) we raise a self-explanatory `TypeError`, showing the trace
that caused it.

Additionally, we introduce a configuration flag
`skip_recursive_messages` (set to `False` by default) that enables us to
avoid the `TypeError` by simply skipping the second occurrence of the
cyclical message. ~The pruning is enabled only for plain messages, and
not for repeated messages or maps. The handling of those cases could be
introduced on a subsequent PR.~